First integration of SANET rest user interface
`SANET <http://sanet.labs.it>`__ has a very nice and responsive restful interface.

I developed SANET when I was employed in Laboratori Guglielmo Marconi,
and I guess that effort that the development team has done since I have left the company
in building this new, restful, interface,
could be reused "as a framework" to build a nice interface for this project.

If so, it could be optimal to factorize out common stuff in a Django reusable app,
we will think about it.

Now, let's try if it is feasible to port the "nui" (they call it this way) they developed
for the sake of our goals.

SANET has been released with GNU Affero Public License (see http://www.gnu.org/licenses/)

NOTE: at the time of this commit I have copied some code, trying to focus only on what we need,
but I cannot do fine-tuning in static/nui/ directory