Model for custom home page definition basing on role and even on user.
* base/views.py: get home page url from rest.models.pages.HomePage
* rest/models.py: moved to rest/models/blocks.py. Not used by gasistafelice
* rest/models/pages.py: added HomePage model and Page model

== HomePage model is simple:

This model bound a role or even a specific user to his default home page.
To do this the model couple (role, user=None) to a resource (his own default resource).

== Page is simple too, but trickier to understand

This model sets default page layout of a resource_type (or even a resource)
for a role, and if you need, a user.

It is the extension of default_settings.RESOURCE_PAGE_BLOCKS

This way you could configure that a GAS Member has "open order" section with "last open order" block.

This is a stub since it is not used in any place now