Refined pages/sections/blocks for basic resources
This is the basic structure of sections and blocks as agreed with
Dominique on 11th of August. This structure is the basic structure which
allow us to present user a basic environment.

Added "order" resource to basic resources to simplify products choice implementation.
So GASSupplierOrder needs full Resource API implementation.