Cancel order on automatic close if no ordered product
This is complete only for automatic close
TODO put signal onto workflow to cath transition in order to execute some code

Cancel on close: If no gasmember order are detected, order is cancelled only if it was not openede the same day.
This exception permit to create fake order, open and close, add extra family curtails and new economic transaction.
For example in some case we need to simullate an order appening in other way (not using this system) but we need to integrate the economic part.