FIX #209 - it "should" be fixed
WARNING: we say "should" because we have been not able to reproduce
exactly the bug. But we have verified that commenting the evaluation of
<script> tags in form popup retrieved, implies that bug (seems) to never
happen.

It could be "bug-prone" if there are <script> tags outside <form> tags
in the html retrieved for building the popup

Anyway we are satified of the work :)