Login has a separate controller and service now
Having a service for the login process let the ui
use it as a component, and it is (should be) so easy to keep
the section active beetween logout and login.

Moreover the login redirection could be implemented such the way
that if any user goes to the #/basket/ url, the ui can open in
the basket section after login.

There are some issues to fix before integrating this work in the
master-dj17 branch:

1. first of all -> FIX the race condition that let no data happen in the
   ui because the main view /person/my does not come back before the
   gasmember/:id/<info> is issued
2. do not use the $location.path for the next variable, but simply
   use the $rootScope.active_section
3. if it is possible fix the "double redirection" after successful login
   you see the browser go to '#/order/' and then to '#/<section>/' and
   it is a bit weird...